Elbow plot

elbow_plot creates an elbow plot.

elbow_plot(dist, maxh = 10)

Arguments

dist
A list object created after running dissimilarity.

maxh
is the maximum "height", or in this case allowable mismatches, as an integer and impacts the length of the x-axis. Default is 10.

Value

It will write to jpg file an elbow plot.

Details

When run it uses the list object created by running dissimilarity and generates an elbow plot. An elbow plot can aid in visualising different "height" parameters and corresponding number of groups created. Here the "height" parameter equates to number of mismatches. The plot is written to the `results/cluster/` sub-directory.

Author

Bart Huntley, bart.huntley@dbca.wa.gov.au

For more details see https://dbca-wa.github.io/ScatMatch/index.html the ScatMatch website

Examples

if (FALSE) {
elbow_plot(dist = dissimilarity_list, maxh = 5)
}