Main function wrapper to calculate genotyping error rates

gen_errors takes a raw data file and generates genotyping error outputs which are written to csv file.

gen_errors(filename, replicates = TRUE, suffix = "_dup")

Arguments

filename
File name of raw data input csv as a character string.

replicates
Logical. TRUE if replicates are present in the raw data.

suffix
File suffix to denote sample replicates. Default is the character string"_dup", change if required. Ignore if replicates = FALSE.

Value

It will write to csv file genotype error file csv's for further processing and/or evaluation.

Details

When run it calculates genotyping errors on the complete raw data set which it reads in from the `source/` sub-directory. The raw data expected is SNP genotypes generated through the Agena Bioscience MassARRAY genotyping system as provided by the Australian Genome Research Facility (AGRF).

Outputs are written to csv file to the `results/` sub-directory as:

numerical_alleles.csv

sample_error_results.csv

summary_error_results.csv

The genotyping error rates (sample amplification rate and, when replicate samples provided, allelic dropout and false allele rates) are used in downstream processes and are further refined.

Author

Bart Huntley, bart.huntley@dbca.wa.gov.au

For more details see https://dbca-wa.github.io/ScatMatch/index.html the ScatMatch website

Examples

if (FALSE) {
gen_errors(filename = "CAGRF20021407_raw.csv", suffix = "_dup")
}