Applies quality filtering of samples by choosing an amplification threshold below which samples are discarded and produces interim filtered results.

amp_threshold takes a threshold value and applies it to the numerical allele data. Also produces a record of which samples were filtered out and proportions of NA's per loci.

amp_threshold(at)

Arguments

at
Numeric value of the average amplification rate to filter the numerical alleles. Values greater than or equal to are retained.

Value

It will write to `results/` three csv files:

filtered alleles data with threshold indicated in the name

sample that were filtered at with threshold indicated in the name

proportion of NA's for the loci

It will also print to screen the number of samples filtered out at this threshold.

Details

The predetermined amplification threshold, ascertained through visualisation amp_splots, is used to filter the numerical alleles data set and written to csv. The names of all the samples that are rejected at this threshold are also written to csv file.

Lastly the proportion of missing data (NA’s) per locus is written to csv. All outputs are written to `results/`.

Author

Bart Huntley, bart.huntley@dbca.wa.gov.au

For more details see https://dbca-wa.github.io/ScatMatch/index.html the ScatMatch website

Examples

if (FALSE) {
amp_threshold(at = 0.8)
}